Move code away from the header
